renamed Item entity `App` namespace to `App\Entity`
