+++ app/src/Entity/Item.php
@@ -1,75 +1,31 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Entity;
 
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity(repositoryClass=ItemRepository::class)
- */
-class Item
+final class Item
 {
     /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
+     * @var string
      */
-    private $id;
+    public $name;
 
     /**
-     * @ORM\Column(type="string", length=255)
+     * @var int
      */
-    private $name;
+    public $sell_in;
 
     /**
-     * @ORM\Column(type="integer")
+     * @var int
      */
-    private $sell_in;
+    public $quality;
 
-    /**
-     * @ORM\Column(type="integer")
-     */
-    private $quality;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(string $name): self
+    public function __construct(string $name, int $sell_in, int $quality)
     {
         $this->name = $name;
-
-        return $this;
-    }
-
-    public function getSellIn(): ?int
-    {
-        return $this->sell_in;
-    }
-
-    public function setSellIn(int $sell_in): self
-    {
         $this->sell_in = $sell_in;
-
-        return $this;
-    }
-
-    public function getQuality(): ?int
-    {
-        return $this->quality;
-    }
-
-    public function setQuality(int $quality): self
-    {
         $this->quality = $quality;
-
-        return $this;
     }
 
     public function __toString(): string
