--- app/src/Repository/StockItemRepository.php
@@ -1,50 +1,0 @@
-<?php
-
-namespace App\Repository;
-
-use App\Entity\StockItem;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-/**
- * @method StockItem|null find($id, $lockMode = null, $lockVersion = null)
- * @method StockItem|null findOneBy(array $criteria, array $orderBy = null)
- * @method StockItem[]    findAll()
- * @method StockItem[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
-class StockItemRepository extends ServiceEntityRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, StockItem::class);
-    }
-
-    // /**
-    //  * @return StockItem[] Returns an array of StockItem objects
-    //  */
-    /*
-    public function findByExampleField($value)
-    {
-        return $this->createQueryBuilder('s')
-            ->andWhere('s.exampleField = :val')
-            ->setParameter('val', $value)
-            ->orderBy('s.id', 'ASC')
-            ->setMaxResults(10)
-            ->getQuery()
-            ->getResult()
-        ;
-    }
-    */
-
-    /*
-    public function findOneBySomeField($value): ?StockItem
-    {
-        return $this->createQueryBuilder('s')
-            ->andWhere('s.exampleField = :val')
-            ->setParameter('val', $value)
-            ->getQuery()
-            ->getOneOrNullResult()
-        ;
-    }
-    */
-}
