+++ app/src/Updater/DefaultUpdater.php
@@ -2,7 +2,7 @@
 
 namespace App\Updater;
 
-use App\Item;
+use App\Entity\Item;
 
 class DefaultUpdater implements UpdaterInterface
 {
