+++ app/tests/Serializer/ItemNormalizerTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Tests\Serializer;
 
-use App\Item;
+use App\Entity\Item;
 use App\Serializer\ItemNormalizer;
 
 use PHPUnit\Framework\TestCase;
