+++ app/tests/Updater/DefaultUpdaterTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Tests\Updater;
 
-use App\Item;
+use App\Entity\Item;
 use App\Updater\DefaultUpdater;
 use PHPUnit\Framework\TestCase;
